Improve ui

+++ app/src/main/java/com/partokarwat/showcase/data/repository/ConversionRateRepository.kt
@@ -1,6 +1,7 @@
 package com.partokarwat.showcase.data.repository
 
 import com.partokarwat.showcase.data.remote.CoinCapApi
+import java.math.BigDecimal
 import javax.inject.Inject
 import javax.inject.Singleton
 
@@ -10,10 +11,10 @@ class ConversionRateRepository
     constructor(
         private val coinCapApi: CoinCapApi,
     ) {
-        suspend fun getExchangeRateToEuro(): Double =
+        suspend fun getExchangeRateToEuro(): BigDecimal =
             coinCapApi
                 .getConversionRateToEUR()
                 .data
                 .rateUsd
-                .toDouble()
+                .toBigDecimal()
     }
